Separate schema change execution components
