+++ task-automator/pkg/phase2/execution.go
@@ -7,7 +7,6 @@ import (
 	"database/sql/driver"
 	"errors"
 	"fmt"
-	"os"
 	"os/exec"
 	"regexp"
 	"strings"
@@ -19,33 +18,16 @@ import (
 	"github.com/shirou/gopsutil/v4/disk"
 )
 
-// BackupMethod represents the type of backup to perform
-type BackupMethod string
-
-const (
-	BackupNone       BackupMethod = "none"
-	BackupMysqldump  BackupMethod = "mysqldump"
-	BackupXtrabackup BackupMethod = "xtrabackup"
-)
-
 var errOnlineDDLUnsupported = errors.New("online DDL is unsupported")
 
 const onlineDDLCleanupTimeout = 5 * time.Second
 
 var (
-	alterOnlineDDLClausePattern = regexp.MustCompile(`(?i)^(?:ALGORITHM|LOCK)\b`)
-	createIndexWaitPattern      = regexp.MustCompile(`(?i)\b(?:WAIT\s+\d+|NOWAIT)\b`)
-	referenceKeywordPattern     = regexp.MustCompile(`(?i)\bREFERENCES\b`)
-	renameOperationPattern      = regexp.MustCompile(`(?i)\bRENAME\b`)
-	renameMemberPattern         = regexp.MustCompile(`(?i)^RENAME\s+(?:COLUMN|INDEX|KEY)\b`)
-	exchangePartitionPattern    = regexp.MustCompile(`(?i)\bEXCHANGE\s+PARTITION\b`)
+	referenceKeywordPattern  = regexp.MustCompile(`(?i)\bREFERENCES\b`)
+	renameOperationPattern   = regexp.MustCompile(`(?i)\bRENAME\b`)
+	renameMemberPattern      = regexp.MustCompile(`(?i)^RENAME\s+(?:COLUMN|INDEX|KEY)\b`)
+	exchangePartitionPattern = regexp.MustCompile(`(?i)\bEXCHANGE\s+PARTITION\b`)
 )
-
-// Executor handles Phase 2 schema change execution
-type Logger interface {
-	Infof(format string, v ...interface{})
-	Errorf(format string, v ...interface{})
-}
 
 type Executor struct {
 	conn       *sql.DB
@@ -71,14 +53,16 @@ func (e *Executor) combinedOutput(name string, args ...string) ([]byte, error) {
 
 // ExecuteOptions contains options for schema change execution
 type ExecuteOptions struct {
-	SQL          string
-	TableName    string
-	Target       *TableInfo
-	BackupMethod BackupMethod
-	OkPTOSC      bool
-	OkOnlineDDL  bool
-	Config       *config.Config // Configuration with paths and directories
-	Debug        bool           // Enable debug output (print commands and outputs)
+	TaskID         int
+	StatementIndex int
+	SQL            string
+	TableName      string
+	Target         *TableInfo
+	BackupMethod   BackupMethod
+	OkPTOSC        bool
+	OkOnlineDDL    bool
+	Config         *config.Config // Configuration with paths and directories
+	Debug          bool           // Enable debug output (print commands and outputs)
 }
 
 // ExecuteResult represents the result of Phase 2 execution
@@ -89,65 +73,6 @@ type ExecuteResult struct {
 	MethodUsed      string
 	Warnings        []string
 	Errors          []string
-}
-
-type executionStage struct {
-	executor *Executor
-	options  ExecuteOptions
-	name     string
-	started  time.Time
-	finished bool
-}
-
-func (e *Executor) beginStage(options ExecuteOptions, name string) *executionStage {
-	stage := &executionStage{executor: e, options: options, name: name, started: time.Now()}
-	e.logStage(options, name, "started", 0, "")
-	return stage
-}
-
-func (s *executionStage) finish(status, reason string) {
-	if s == nil || s.finished {
-		return
-	}
-	s.finished = true
-	s.executor.logStage(s.options, s.name, status, time.Since(s.started), reason)
-}
-
-func (e *Executor) logStage(options ExecuteOptions, stage, status string, duration time.Duration, reason string) {
-	if e.logger == nil {
-		return
-	}
-	table := executionLogTable(options)
-	message := fmt.Sprintf("Schema change check: stage=%s status=%s table=%s", stage, status, table)
-	if status != "started" {
-		message += fmt.Sprintf(" duration_ms=%d", duration.Milliseconds())
-	}
-	if reason = sanitizeExecutionLogReason(options, reason); reason != "" {
-		message += fmt.Sprintf(" reason=%q", reason)
-	}
-	if status == "failed" {
-		e.logger.Errorf("%s", message)
-		return
-	}
-	e.logger.Infof("%s", message)
-}
-
-func executionLogTable(options ExecuteOptions) string {
-	if options.Target != nil && options.Target.Database != "" && options.Target.Table != "" {
-		return options.Target.Database + "." + options.Target.Table
-	}
-	if table := strings.Trim(options.TableName, " `"); table != "" {
-		return strings.ReplaceAll(table, "`.`", ".")
-	}
-	return "unknown"
-}
-
-func sanitizeExecutionLogReason(options ExecuteOptions, reason string) string {
-	reason = strings.Join(strings.Fields(reason), " ")
-	if options.Config != nil && options.Config.MysqlPassword != "" {
-		reason = strings.ReplaceAll(reason, options.Config.MysqlPassword, "***")
-	}
-	return reason
 }
 
 // Execute performs Phase 2 schema change execution
@@ -217,18 +142,15 @@ func (e *Executor) Execute(options ExecuteOptions) (*ExecuteResult, error) {
 	if options.OkOnlineDDL {
 		methodStage := e.beginStage(options, "online_ddl")
 		if err := e.executeWithOnlineDDL(options, result); err != nil {
-			methodStage.finish("failed", err.Error())
+			methodStage.finish("failed", executionStageFailureReason(err))
 			if options.OkPTOSC && isOnlineDDLUnsupported(err) {
 				fallbackStage := e.beginStage(options, "ptosc_fallback")
-				fallbackStage.finish("passed", err.Error())
+				fallbackStage.finish("passed", "trigger=online_ddl_unsupported")
 				result.Warnings = append(result.Warnings,
 					fmt.Sprintf("Online DDL unavailable; used pt-online-schema-change: %v", err))
-				if e.logger != nil {
-					e.logger.Infof("Online DDL not possible for table %s (%v); falling back to pt-online-schema-change", options.TableName, err)
-				}
 				ptoscStage := e.beginStage(options, "ptosc")
 				if ptErr := e.executeWithPTOSC(options); ptErr != nil {
-					ptoscStage.finish("failed", ptErr.Error())
+					ptoscStage.finish("failed", executionStageFailureReason(ptErr))
 					return nil, fmt.Errorf("pt-online-schema-change execution failed: %w", ptErr)
 				}
 				ptoscStage.finish("passed", "fallback completed")
@@ -252,7 +174,7 @@ func (e *Executor) Execute(options ExecuteOptions) (*ExecuteResult, error) {
 		stage.finish("skipped", "Platform analysis did not allow Online DDL")
 		ptoscStage := e.beginStage(options, "ptosc")
 		if err := e.executeWithPTOSC(options); err != nil {
-			ptoscStage.finish("failed", err.Error())
+			ptoscStage.finish("failed", executionStageFailureReason(err))
 			return nil, fmt.Errorf("pt-online-schema-change execution failed: %w", err)
 		}
 		ptoscStage.finish("passed", "")
@@ -442,58 +364,6 @@ func (e *Executor) tableTargetsEqual(first, second TableInfo) (bool, error) {
 	return lowerCaseTableNames == 1 || lowerCaseTableNames == 2, nil
 }
 
-func (e *Executor) performBackup(options ExecuteOptions) (string, error) {
-	if options.Config == nil {
-		return "", fmt.Errorf("config is required for backup")
-	}
-	filesystemStage := e.beginStage(options, "backup_filesystem")
-	usage, err := prepareBackupFilesystem(options.Config.BackupDir, !options.Config.DisableSpaceChecks)
-	if err != nil {
-		filesystemStage.finish("failed", err.Error())
-		return "", err
-	}
-	filesystemStage.finish("passed", "backup directory is ready")
-
-	// Check disk space before performing backup
-	if !options.Config.DisableSpaceChecks {
-		capacityStage := e.beginStage(options, "backup_capacity")
-		if err := e.checkDiskSpace(options, usage); err != nil {
-			capacityStage.finish("failed", err.Error())
-			if e.logger != nil {
-				e.logger.Errorf("disk space check failed for table %s: %v", options.TableName, err)
-			}
-			return "", err
-		}
-		capacityStage.finish("passed", "")
-	} else {
-		stage := e.beginStage(options, "backup_capacity")
-		stage.finish("skipped", "space checks are disabled")
-	}
-
-	switch options.BackupMethod {
-	case BackupMysqldump:
-		return e.backupWithMysqldump(options)
-	case BackupXtrabackup:
-		return e.backupWithXtrabackup(options)
-	default:
-		return "", fmt.Errorf("unsupported backup method: %s", options.BackupMethod)
-	}
-}
-
-func prepareBackupFilesystem(path string, inspectUsage bool) (*disk.UsageStat, error) {
-	if err := os.MkdirAll(path, 0755); err != nil {
-		return nil, fmt.Errorf("failed to create backup directory: %w", err)
-	}
-	if !inspectUsage {
-		return nil, nil
-	}
-	usage, err := disk.Usage(path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to check disk space: %w", err)
-	}
-	return usage, nil
-}
-
 func (e *Executor) checkDataDirFilesystemCapacity(options ExecuteOptions) error {
 	tableInfo, err := resolvedExecutionTarget(options)
 	if err != nil {
@@ -558,616 +428,6 @@ func (e *Executor) checkDataDirFilesystemCapacity(options ExecuteOptions) error 
 	}
 
 	return nil
-}
-
-func (e *Executor) backupWithMysqldump(options ExecuteOptions) (string, error) {
-	if options.Config == nil {
-		return "", fmt.Errorf("config is required for backup")
-	}
-
-	host := options.Config.MysqlHost
-	port := options.Config.MysqlPort
-	user := options.Config.MysqlUser
-	password := options.Config.MysqlPassword
-	if host == "" {
-		return "", fmt.Errorf("mysql_host is required for backup")
-	}
-
-	tableInfo, err := resolvedExecutionTarget(options)
-	if err != nil {
-		return "", err
-	}
-
-	// Use config values
-	mysqldump := options.Config.MysqldumpPath
-	if mysqldump == "" {
-		mysqldump = "mysqldump"
-	}
-
-	// Generate timestamp prefix in YYMMDDHHMMSS format
-	timestamp := time.Now().Format("060102150405")
-	backupPath := fmt.Sprintf("%s/%s_%s_%s.sql", options.Config.BackupDir, timestamp, tableInfo.Database, tableInfo.Table)
-
-	args := buildMysqldumpConnectionArgs(host, port, user, password)
-	args = append(args,
-		tableInfo.Database,
-		tableInfo.Table,
-		"--single-transaction",
-		"--quick",
-		"--lock-tables=false",
-		"-r", backupPath,
-	)
-
-	cmd := exec.Command(mysqldump, args...)
-
-	if options.Debug {
-		// Mask password in debug output
-		safeArgs := make([]string, len(args))
-		copy(safeArgs, args)
-		// Find and mask password argument (-p followed by password)
-		for i, arg := range safeArgs {
-			if strings.HasPrefix(arg, "-p") && len(arg) > 2 {
-				safeArgs[i] = "-p***"
-			}
-		}
-		e.debugf(options, "[DEBUG] mysqldump command: %s", mysqldump)
-		e.debugf(options, "[DEBUG] mysqldump args: %s", strings.Join(safeArgs, " "))
-	}
-
-	output, err := cmd.CombinedOutput()
-	if options.Debug {
-		if len(output) > 0 {
-			e.debugf(options, "[DEBUG] mysqldump output:\n%s", string(output))
-		}
-	}
-
-	if err != nil {
-		if options.Debug {
-			e.debugf(options, "[DEBUG] mysqldump error: %v", err)
-		}
-		return "", fmt.Errorf("mysqldump failed: %w", err)
-	}
-
-	return backupPath, nil
-}
-
-func buildMysqldumpConnectionArgs(host, port, user, password string) []string {
-	args := []string{"-u", user, "-p" + password}
-	if strings.HasPrefix(host, "/") {
-		return append([]string{"--socket=" + host}, args...)
-	}
-	return append([]string{"-h", host, "-P", port}, args...)
-}
-
-func buildXtrabackupConnectionArgs(host, port, user, password string) []string {
-	args := []string{
-		"--user=" + user,
-		"--password=" + password,
-	}
-	if strings.HasPrefix(host, "/") {
-		return append(args, "--socket="+host)
-	}
-	return append(args, "--host="+host, "--port="+port)
-}
-
-func (e *Executor) backupWithXtrabackup(options ExecuteOptions) (string, error) {
-	if options.Config == nil {
-		return "", fmt.Errorf("config is required for backup")
-	}
-
-	host := options.Config.MysqlHost
-	port := options.Config.MysqlPort
-	user := options.Config.MysqlUser
-	password := options.Config.MysqlPassword
-	if host == "" {
-		return "", fmt.Errorf("mysql_host is required for backup")
-	}
-
-	tableInfo, err := resolvedExecutionTarget(options)
-	if err != nil {
-		return "", err
-	}
-
-	// Use config values
-	xtrabackup := options.Config.XtrabackupPath
-	if xtrabackup == "" {
-		xtrabackup = "xtrabackup"
-	}
-
-	// Generate timestamp prefix in YYMMDDHHMMSS format
-	timestamp := time.Now().Format("060102150405")
-	// Create a unique backup directory for this table
-	backupDir := fmt.Sprintf("%s/%s_xtrabackup_%s_%s", options.Config.BackupDir, timestamp, tableInfo.Database, tableInfo.Table)
-
-	// Step 1: Take backup of the table using --tables option.
-	// xtrabackup treats --tables as a regex, so escape metacharacters and anchor
-	// both sides to match only the exact database.table name.
-	tableName := fmt.Sprintf("%s.%s", tableInfo.Database, tableInfo.Table)
-	tableSpec := "^" + regexp.QuoteMeta(tableName) + "$"
-
-	backupArgs := []string{
-		"--backup",
-		"--ftwrl-wait-timeout=15",
-		"--tables=" + tableSpec,
-		"--target-dir=" + backupDir,
-	}
-	backupArgs = append(backupArgs, buildXtrabackupConnectionArgs(host, port, user, password)...)
-
-	if options.Debug {
-		// Mask password in debug output
-		safeArgs := make([]string, len(backupArgs))
-		copy(safeArgs, backupArgs)
-		for i, arg := range safeArgs {
-			if strings.HasPrefix(arg, "--password=") {
-				safeArgs[i] = "--password=***"
-			}
-		}
-		e.debugf(options, "[DEBUG] xtrabackup backup command: %s", xtrabackup)
-		e.debugf(options, "[DEBUG] xtrabackup backup args: %s", strings.Join(safeArgs, " "))
-	}
-
-	cmd := exec.Command(xtrabackup, backupArgs...)
-	output, err := cmd.CombinedOutput()
-	if options.Debug {
-		if len(output) > 0 {
-			e.debugf(options, "[DEBUG] xtrabackup backup output:\n%s", string(output))
-		}
-		if err != nil {
-			e.debugf(options, "[DEBUG] xtrabackup backup error: %v", err)
-		}
-	}
-
-	if err != nil {
-		return "", fmt.Errorf("xtrabackup backup failed: %w: %s", err, string(output))
-	}
-
-	// Step 2: Prepare the backup with --export option
-	// This prepares the backup and exports table metadata for transportable tablespace
-	prepareArgs := []string{
-		"--prepare",
-		"--export",
-		"--target-dir=" + backupDir,
-	}
-
-	if options.Debug {
-		e.debugf(options, "[DEBUG] xtrabackup prepare command: %s", xtrabackup)
-		e.debugf(options, "[DEBUG] xtrabackup prepare args: %s", strings.Join(prepareArgs, " "))
-	}
-
-	cmd = exec.Command(xtrabackup, prepareArgs...)
-	output, err = cmd.CombinedOutput()
-	if options.Debug {
-		if len(output) > 0 {
-			e.debugf(options, "[DEBUG] xtrabackup prepare output:\n%s", string(output))
-		}
-		if err != nil {
-			e.debugf(options, "[DEBUG] xtrabackup prepare error: %v", err)
-		}
-	}
-
-	if err != nil {
-		return "", fmt.Errorf("xtrabackup prepare failed: %w: %s", err, string(output))
-	}
-
-	// Return the backup directory path
-	// The table files (.ibd and .cfg) will be in backupDir/database/table.*
-	return backupDir, nil
-}
-
-// checkDiskSpace checks if there's enough disk space for the backup
-func (e *Executor) checkDiskSpace(options ExecuteOptions, usage *disk.UsageStat) error {
-	var estimatedSize int64
-
-	tableInfo, err := resolvedExecutionTarget(options)
-	if err != nil {
-		return fmt.Errorf("failed to parse table name: %w", err)
-	}
-
-	// Estimate backup size based on method
-	switch options.BackupMethod {
-	case BackupMysqldump:
-		sizeMB, err := e.estimateMysqldumpSize(tableInfo.Database, tableInfo.Table)
-		if err != nil {
-			return fmt.Errorf("failed to estimate backup size: %w", err)
-		}
-		estimatedSize = int64(sizeMB * 1024 * 1024) // Convert MB to bytes
-	case BackupXtrabackup:
-		// Xtrabackup backs up entire database, so we estimate based on database size
-		sizeMB, err := e.estimateXtrabackupSize(tableInfo.Database)
-		if err != nil {
-			return fmt.Errorf("failed to estimate backup size: %w", err)
-		}
-		estimatedSize = int64(sizeMB * 1024 * 1024) // Convert MB to bytes
-	default:
-		return nil // No backup, no space check needed
-	}
-
-	// Add buffer percentage to estimated size
-	bufferPercent := options.Config.BackupSpaceBuffer
-	if bufferPercent == 0 {
-		bufferPercent = 20.0 // Fallback to 20% if not configured
-	}
-	requiredSize := estimatedSize + int64(float64(estimatedSize)*bufferPercent/100.0)
-
-	if usage == nil {
-		return fmt.Errorf("backup filesystem usage is required for disk space check")
-	}
-
-	availableBytes := int64(usage.Free)
-
-	if options.Debug {
-		e.debugf(options, "[DEBUG] Estimated backup size: %.2f MB", float64(estimatedSize)/(1024*1024))
-		e.debugf(options, "[DEBUG] Required space (with %.1f%% buffer): %.2f MB", bufferPercent, float64(requiredSize)/(1024*1024))
-		e.debugf(options, "[DEBUG] Available disk space: %.2f MB", float64(availableBytes)/(1024*1024))
-	}
-
-	if availableBytes < requiredSize {
-		return fmt.Errorf("insufficient disk space: required %.2f MB (with %.1f%% buffer), available %.2f MB",
-			float64(requiredSize)/(1024*1024),
-			bufferPercent,
-			float64(availableBytes)/(1024*1024))
-	}
-
-	return nil
-}
-
-// estimateMysqldumpSize estimates the size of a mysqldump backup for a specific table
-func (e *Executor) estimateMysqldumpSize(dbName, tableName string) (float64, error) {
-	var dataLength, indexLength sql.NullInt64
-	err := e.conn.QueryRow(`
-		SELECT DATA_LENGTH, INDEX_LENGTH
-		FROM information_schema.TABLES
-		WHERE TABLE_SCHEMA = ? AND TABLE_NAME = ?
-	`, dbName, tableName).Scan(&dataLength, &indexLength)
-	if err != nil {
-		return 0, err
-	}
-
-	var totalBytes int64
-	if dataLength.Valid {
-		totalBytes += dataLength.Int64
-	}
-	if indexLength.Valid {
-		totalBytes += indexLength.Int64
-	}
-
-	if totalBytes == 0 {
-		return 0.1, nil // Return a minimal size estimate if table size is 0 or NULL
-	}
-
-	// mysqldump typically produces 1.5-2x the table size due to SQL format overhead
-	estimatedSizeMB := float64(totalBytes) * 2.0 / (1024 * 1024)
-
-	return estimatedSizeMB, nil
-}
-
-// estimateXtrabackupSize estimates the size of an xtrabackup for the entire database
-func (e *Executor) estimateXtrabackupSize(dbName string) (float64, error) {
-	var totalBytes sql.NullInt64
-	err := e.conn.QueryRow(`
-		SELECT SUM(DATA_LENGTH + INDEX_LENGTH)
-		FROM information_schema.TABLES
-		WHERE TABLE_SCHEMA = ?
-	`, dbName).Scan(&totalBytes)
-	if err != nil {
-		return 0, err
-	}
-
-	if !totalBytes.Valid || totalBytes.Int64 == 0 {
-		return 0.1, nil // Return a minimal size estimate if database size is 0 or NULL
-	}
-
-	// Xtrabackup includes all tables, indexes, and some overhead
-	// Estimate ~1.2x the database size
-	estimatedSizeMB := float64(totalBytes.Int64) * 1.2 / (1024 * 1024)
-
-	return estimatedSizeMB, nil
-}
-
-func (e *Executor) executeWithPTOSC(options ExecuteOptions) error {
-	// Perform dry-run first
-	dryRunStage := e.beginStage(options, "ptosc_dry_run")
-	if err := e.dryRunPTOSC(options); err != nil {
-		dryRunStage.finish("failed", err.Error())
-		return fmt.Errorf("pt-online-schema-change dry-run failed: %w", err)
-	}
-	dryRunStage.finish("passed", "")
-
-	// Execute actual change
-	executionStage := e.beginStage(options, "ptosc_execution")
-	err := e.runPTOSC(options)
-	if err != nil {
-		executionStage.finish("failed", err.Error())
-		return err
-	}
-	executionStage.finish("passed", "")
-	return nil
-}
-
-func (e *Executor) dryRunPTOSC(options ExecuteOptions) error {
-	if options.Config == nil {
-		return fmt.Errorf("config is required for pt-online-schema-change")
-	}
-
-	ptosc := options.Config.PTOSCPath
-	if ptosc == "" {
-		ptosc = "pt-online-schema-change"
-	}
-
-	host := options.Config.MysqlHost
-	port := options.Config.MysqlPort
-	user := options.Config.MysqlUser
-	password := options.Config.MysqlPassword
-	if host == "" {
-		return fmt.Errorf("mysql_host is required for pt-online-schema-change")
-	}
-
-	tableInfo, err := resolvedExecutionTarget(options)
-	if err != nil {
-		return err
-	}
-
-	alterSQL, err := buildPTOSCAlterSQL(options.SQL)
-	if err != nil {
-		return err
-	}
-
-	args := []string{
-		"--dry-run",
-		buildPTOSCDSN(host, port, user, password, tableInfo.Database, tableInfo.Table),
-		fmt.Sprintf("--alter=%s", alterSQL),
-	}
-
-	if options.Debug {
-		// Mask password in debug output
-		safeArgs := make([]string, len(args))
-		copy(safeArgs, args)
-		for i, arg := range safeArgs {
-			// Mask password in connection string (p=password)
-			if strings.Contains(arg, ",p=") {
-				parts := strings.Split(arg, ",p=")
-				if len(parts) == 2 {
-					// Extract everything after p= and before next comma
-					rest := strings.Split(parts[1], ",")
-					if len(rest) > 0 {
-						safeArgs[i] = parts[0] + ",p=***," + strings.Join(rest[1:], ",")
-					} else {
-						safeArgs[i] = parts[0] + ",p=***"
-					}
-				} else if strings.HasPrefix(arg, "p=") {
-					rest := strings.SplitN(arg, ",", 2)
-					safeArgs[i] = "p=***"
-					if len(rest) > 1 {
-						safeArgs[i] += "," + rest[1]
-					}
-				}
-			}
-		}
-		e.debugf(options, "[DEBUG] pt-online-schema-change dry-run command: %s", ptosc)
-		e.debugf(options, "[DEBUG] pt-online-schema-change dry-run args: %s", strings.Join(safeArgs, " "))
-	}
-
-	output, err := e.combinedOutput(ptosc, args...)
-	e.debugf(options, "[DEBUG] pt-online-schema-change dry-run output:\n%s", string(output))
-	if err != nil {
-		e.debugf(options, "[DEBUG] pt-online-schema-change dry-run error: %v", err)
-	}
-
-	if err != nil {
-		return fmt.Errorf("pt-online-schema-change dry-run failed: %s", string(output))
-	}
-
-	return nil
-}
-
-func (e *Executor) runPTOSC(options ExecuteOptions) error {
-	if options.Config == nil {
-		return fmt.Errorf("config is required for pt-online-schema-change")
-	}
-
-	ptosc := options.Config.PTOSCPath
-	if ptosc == "" {
-		ptosc = "pt-online-schema-change"
-	}
-
-	host := options.Config.MysqlHost
-	port := options.Config.MysqlPort
-	user := options.Config.MysqlUser
-	password := options.Config.MysqlPassword
-	if host == "" {
-		return fmt.Errorf("mysql_host is required for pt-online-schema-change")
-	}
-
-	tableInfo, err := resolvedExecutionTarget(options)
-	if err != nil {
-		return err
-	}
-
-	alterSQL, err := buildPTOSCAlterSQL(options.SQL)
-	if err != nil {
-		return err
-	}
-
-	args := []string{
-		"--execute",
-		buildPTOSCDSN(host, port, user, password, tableInfo.Database, tableInfo.Table),
-		fmt.Sprintf("--alter=%s", alterSQL),
-	}
-
-	if options.Debug {
-		// Mask password in debug output
-		safeArgs := make([]string, len(args))
-		copy(safeArgs, args)
-		for i, arg := range safeArgs {
-			// Mask password in connection string (p=password)
-			if strings.Contains(arg, ",p=") {
-				parts := strings.Split(arg, ",p=")
-				if len(parts) == 2 {
-					// Extract everything after p= and before next comma
-					rest := strings.Split(parts[1], ",")
-					if len(rest) > 0 {
-						safeArgs[i] = parts[0] + ",p=***," + strings.Join(rest[1:], ",")
-					} else {
-						safeArgs[i] = parts[0] + ",p=***"
-					}
-				} else if strings.HasPrefix(arg, "p=") {
-					rest := strings.SplitN(arg, ",", 2)
-					safeArgs[i] = "p=***"
-					if len(rest) > 1 {
-						safeArgs[i] += "," + rest[1]
-					}
-				}
-			}
-		}
-		e.debugf(options, "[DEBUG] pt-online-schema-change execute command: %s", ptosc)
-		e.debugf(options, "[DEBUG] pt-online-schema-change execute args: %s", strings.Join(safeArgs, " "))
-	}
-
-	output, err := e.combinedOutput(ptosc, args...)
-	e.debugf(options, "[DEBUG] pt-online-schema-change execute output:\n%s", string(output))
-	if err != nil {
-		e.debugf(options, "[DEBUG] pt-online-schema-change execute error: %v", err)
-	}
-
-	if err != nil {
-		return fmt.Errorf("pt-online-schema-change failed: %s", string(output))
-	}
-
-	return nil
-}
-
-func buildPTOSCDSN(host, port, user, password, database, table string) string {
-	parts := make([]string, 0, 6)
-	if strings.HasPrefix(host, "/") {
-		parts = append(parts, "S="+host)
-	} else {
-		parts = append(parts, "h="+host, "P="+port)
-	}
-	parts = append(parts, "u="+user, "p="+password, "D="+database, "t="+table)
-	return strings.Join(parts, ",")
-}
-
-func buildPTOSCAlterSQL(sql string) (string, error) {
-	if containsExecutableSQLComment(sql) {
-		return "", fmt.Errorf("executable SQL comments are not allowed in pt-online-schema-change input")
-	}
-	if containsAmbiguousBackslashQuote(sql) {
-		return "", fmt.Errorf("backslash-escaped quotes are not allowed in pt-online-schema-change input because their meaning depends on sql_mode")
-	}
-
-	statement, _ := splitDDLStatementSuffix(sql)
-	if statement == "" {
-		return "", fmt.Errorf("empty SQL statement")
-	}
-	if strings.Contains(maskSQLStringsAndComments(statement, true), ";") {
-		return "", fmt.Errorf("multiple SQL statements are not allowed in pt-online-schema-change input")
-	}
-
-	masked := maskSQLStringsAndComments(statement, false)
-	if match := alterTableTargetPattern.FindStringSubmatchIndex(masked); len(match) >= 4 {
-		modifiers := strings.Fields(masked[:match[2]])
-		for i, modifier := range modifiers {
-			if strings.EqualFold(modifier, "IGNORE") {
-				return "", fmt.Errorf("ALTER IGNORE cannot be represented safely by pt-online-schema-change")
-			}
-			if strings.EqualFold(modifier, "IF") && i+1 < len(modifiers) && strings.EqualFold(modifiers[i+1], "EXISTS") {
-				return "", fmt.Errorf("ALTER TABLE IF EXISTS cannot be represented safely by pt-online-schema-change")
-			}
-		}
-		return stripAlterOnlineDDLClauses(statement[match[3]:])
-	}
-
-	match := createIndexPTOSCPattern.FindStringSubmatchIndex(masked)
-	if len(match) < 6 {
-		return "", fmt.Errorf("unsupported DDL for pt-online-schema-change; expected ALTER TABLE or CREATE INDEX")
-	}
-	definition := strings.TrimSpace(statement[match[2]:match[3]])
-	tail := stripCreateIndexOnlineDDLClauses(statement[match[5]:])
-	if tail == "" || !strings.HasPrefix(strings.TrimSpace(tail), "(") {
-		return "", fmt.Errorf("invalid CREATE INDEX column definition")
-	}
-	if createIndexWaitPattern.MatchString(maskTopLevelSQL(tail)) {
-		return "", fmt.Errorf("CREATE INDEX WAIT/NOWAIT cannot be represented safely by pt-online-schema-change")
-	}
-	return "ADD " + definition + " " + strings.TrimSpace(tail), nil
-}
-
-func stripAlterOnlineDDLClauses(body string) (string, error) {
-	masked := maskSQLStringsAndComments(body, true)
-	start := 0
-	depth := 0
-	kept := make([]string, 0, 4)
-
-	appendClause := func(end int) {
-		rawClause := strings.TrimSpace(body[start:end])
-		maskedClause := strings.TrimSpace(masked[start:end])
-		if rawClause != "" && !alterOnlineDDLClausePattern.MatchString(maskedClause) {
-			kept = append(kept, rawClause)
-		}
-	}
-
-	for i := range masked {
-		switch masked[i] {
-		case '(':
-			depth++
-		case ')':
-			if depth > 0 {
-				depth--
-			}
-		case ',':
-			if depth == 0 {
-				appendClause(i)
-				start = i + 1
-			}
-		}
-	}
-	appendClause(len(body))
-
-	if len(kept) == 0 {
-		return "", fmt.Errorf("ALTER TABLE contains no change supported by pt-online-schema-change")
-	}
-	return strings.Join(kept, ", "), nil
-}
-
-func stripCreateIndexOnlineDDLClauses(tail string) string {
-	masked := []byte(maskTopLevelSQL(tail))
-
-	clausePattern := regexp.MustCompile(`(?i)\b(?:ALGORITHM|LOCK)\b\s*(?:=\s*|\s+)(?:DEFAULT|INPLACE|COPY|INSTANT|NOCOPY|NONE|SHARED|EXCLUSIVE)\b`)
-	spans := clausePattern.FindAllStringIndex(string(masked), -1)
-	if len(spans) == 0 {
-		return strings.TrimSpace(tail)
-	}
-
-	var result strings.Builder
-	last := 0
-	for _, span := range spans {
-		result.WriteString(tail[last:span[0]])
-		last = span[1]
-	}
-	result.WriteString(tail[last:])
-	return strings.TrimSpace(result.String())
-}
-
-func maskTopLevelSQL(sql string) string {
-	masked := []byte(maskSQLStringsAndComments(sql, true))
-	depth := 0
-	for i := range masked {
-		switch masked[i] {
-		case '(':
-			depth++
-			masked[i] = ' '
-		case ')':
-			if depth > 0 {
-				depth--
-			}
-			masked[i] = ' '
-		default:
-			if depth > 0 {
-				masked[i] = ' '
-			}
-		}
-	}
-	return string(masked)
 }
 
 func isOnlineDDLUnsupported(err error) bool {
@@ -1241,14 +501,20 @@ func (e *Executor) executeWithOnlineDDL(options ExecuteOptions, result *ExecuteR
 			schemaStage := e.beginStage(options, "online_ddl_scratch_schema")
 			if _, callbackErr = conn.ExecContext(ctx, fmt.Sprintf("CREATE DATABASE IF NOT EXISTS `%s`", escapeIdent(testSchema))); callbackErr != nil {
 				schemaStage.finish("failed", callbackErr.Error())
-				return fmt.Errorf("failed to create test schema %s: %w", testSchema, callbackErr)
+				return wrapExecutionStageError(
+					"online_ddl_scratch_schema",
+					fmt.Errorf("failed to create test schema %s: %w", testSchema, callbackErr),
+				)
 			}
 			schemaStage.finish("passed", "")
 
 			tableStage := e.beginStage(options, "online_ddl_scratch_table")
 			if _, callbackErr = conn.ExecContext(ctx, fmt.Sprintf("CREATE TABLE %s LIKE %s", testTableRef, sourceTableRef)); callbackErr != nil {
 				tableStage.finish("failed", callbackErr.Error())
-				return fmt.Errorf("failed to create test table %s: %w", testTableRef, callbackErr)
+				return wrapExecutionStageError(
+					"online_ddl_scratch_table",
+					fmt.Errorf("failed to create test table %s: %w", testTableRef, callbackErr),
+				)
 			}
 			tableStage.finish("passed", "")
 			defer func() {
@@ -1265,7 +531,10 @@ func (e *Executor) executeWithOnlineDDL(options ExecuteOptions, result *ExecuteR
 			preflightStage := e.beginStage(options, "online_ddl_preflight")
 			if _, callbackErr = conn.ExecContext(ctx, testSQL); callbackErr != nil {
 				preflightStage.finish("failed", callbackErr.Error())
-				return fmt.Errorf("online DDL preflight failed on test table %s: %w", testTableRef, callbackErr)
+				return wrapExecutionStageError(
+					"online_ddl_preflight",
+					fmt.Errorf("online DDL preflight failed on test table %s: %w", testTableRef, callbackErr),
+				)
 			}
 			preflightStage.finish("passed", "")
 
@@ -1274,14 +543,15 @@ func (e *Executor) executeWithOnlineDDL(options ExecuteOptions, result *ExecuteR
 			_, callbackErr = conn.ExecContext(ctx, finalSQL)
 			if callbackErr != nil {
 				productionStage.finish("failed", callbackErr.Error())
+				return wrapExecutionStageError("online_ddl_production", callbackErr)
 			} else {
 				productionStage.finish("passed", "")
 			}
-			return callbackErr
+			return nil
 		},
 	)
 	if executionErr != nil {
-		sessionStage.finish("failed", executionErr.Error())
+		sessionStage.finish("failed", executionStageFailureReason(executionErr))
 	} else if cleanupErr != nil {
 		sessionStage.finish("failed", cleanupErr.Error())
 	} else {
@@ -1302,7 +572,6 @@ func (e *Executor) executeWithOnlineDDL(options ExecuteOptions, result *ExecuteR
 	err = executionErr
 	if err != nil {
 		result.Warnings = append(result.Warnings, err.Error())
-		e.debugf(options, "[DEBUG] Online DDL execution error: %v", err)
 	} else {
 		e.debugf(options, "[DEBUG] Online DDL execution successful")
 	}
@@ -1322,13 +591,6 @@ func (e *Executor) executeWithOnlineDDL(options ExecuteOptions, result *ExecuteR
 func buildOnlineDDLTestTableName(schema, table string, nonce int64) string {
 	digest := sha256.Sum256([]byte(fmt.Sprintf("%s\x00%s\x00%d", schema, table, nonce)))
 	return fmt.Sprintf("_releem_ddl_test_%x", digest[:16])
-}
-
-func executeOnlineDDLOnPinnedConn(ctx context.Context, db *sql.DB, ddl string, lockWaitTimeout int64) (executionErr, cleanupErr error) {
-	return withPinnedLockWaitTimeout(ctx, db, lockWaitTimeout, func(ctx context.Context, conn *sql.Conn) error {
-		_, err := conn.ExecContext(ctx, ddl)
-		return err
-	})
 }
 
 func withPinnedLockWaitTimeout(
@@ -1563,15 +825,4 @@ func rewriteDDLTargetTable(sql, newTableRef string) (string, error) {
 
 func escapeIdent(id string) string {
 	return strings.ReplaceAll(id, "`", "``")
-}
-
-func (e *Executor) debugf(options ExecuteOptions, format string, args ...interface{}) {
-	if !options.Debug {
-		return
-	}
-	if e.logger != nil {
-		e.logger.Infof(format, args...)
-		return
-	}
-	fmt.Printf(format+"\n", args...)
 }
